backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
openai==1.58.1
chromadb==0.5.23
python-dotenv==1.0.1
pydantic==2.10.4
